Add files via upload
we have been assigned to make an integrated project and we have chosen the name healthsphere, for our project, as working on it. this is the final prototype of our website. which includes the advanced and improved version of the previous code, and has the readings based on the dummy data.